infra/n8n/autoscaler/requirements.txt
redis==5.0.7
docker==7.1.0
python-dotenv==1.0.1

